Move tests from `src/compute/` to `tests/`
