+++ tests/it/compute/aggregate/mod.rs
@@ -1,0 +1,3 @@
+mod memory;
+mod min_max;
+mod sum;
