+++ tests/it/compute/arithmetics/basic/mod.rs
@@ -1,0 +1,6 @@
+mod add;
+mod div;
+mod mul;
+mod pow;
+mod rem;
+mod sub;
