+++ tests/it/compute/arithmetics/decimal/mod.rs
@@ -1,0 +1,4 @@
+mod add;
+mod div;
+mod mul;
+mod sub;
